refactor : 노트메뉴, 노트 음악 실행 부분을 컴포넌트화 하여 재사용성 ↑
- 노트 검색 상세화면 VC 그리고 노트 및 댓글화면 VC에서도 동일한 기능이 필요로 한 상황에서
프로토콜 기본구현을 활용하여 노트 메뉴 그리고 노트 음악 실행 부분을 컴포넌트화 하여 중복 코드 방지
